requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.6
Django==4.0.1
django-allauth==0.50.0
django-autocomplete-light==3.9.4
django-autoslug==1.9.8
django-bootstrap-modal-forms==2.2.0
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-email-verification==0.2.2
django-environ==0.8.1
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.0.0
django-livereload-server==0.4
django-taggit==2.1.0
django-taggit-autosuggest==0.3.8
django-tags-input==5.1.0
django-tinymce==3.4.0
django-utils-six==2.0
Django-Verify-Email==2.0.3
django-widget-tweaks==1.4.12
gunicorn==20.1.0
lazy-object-proxy==1.4.3
matplotlib==3.3.3
numpy==1.23.3
oauthlib==3.2.0
opencv-python==4.6.0.66
pandas==1.5.0
Pillow==9.2.0
psycopg2==2.9.5
reportlab==3.6.10
requests==2.27.1
requests-oauthlib==1.3.1
virtualenv==20.1.0
whitenoise==6.2.0
